move results into codegen models
