+++ src/seer/automation/models.py
@@ -743,13 +743,3 @@ class FileChange(BaseModel):
             return None
 
         return file_contents.replace(self.reference_snippet, "")
-
-
-class RelevantWarningResult(BaseModel):
-    warning_id: int
-    issue_group_id: str
-    does_fixing_warning_fix_issue: bool
-    relevance_probability: float
-    reasoning: str
-    short_description: str
-    short_justification: str
